A cable retainer for mounting inside cabinet
This holds onto a USB-C cable that runs inside a cabinet. When the cable
is let go, it drops into this and stay there for easy grasping.

It's intended to be attached with double-sided tape (bathroom mirror
mounting stuff) attached to the max-z surface of retainer(), with the
hole protruding. Obviously, the cable has to go in before an adhesive is
used.

Using cutout() is kind of optional.